sdk/python/pyproject.toml
[project]
name = "auth-platform-sdk"
version = "1.0.0"
description = "Auth Platform Python SDK - December 2025 State of Art"
authors = [{name = "Auth Platform Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Framework :: AsyncIO",
]
keywords = ["auth", "oauth", "jwt", "dpop", "pkce", "sdk"]
dependencies = [
    "httpx>=0.28.0",
    "pyjwt[crypto]>=2.10.0",
    "pydantic>=2.10.0",
    "cryptography>=44.0.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.120.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "respx>=0.22.0",
    "coverage>=7.6.0",
]
fastapi = ["fastapi>=0.115.0"]
flask = ["flask>=3.1.0"]
django = ["django>=5.1.0"]

[project.urls]
Homepage = "https://github.com/auth-platform/python-sdk"
Documentation = "https://docs.auth-platform.io/sdk/python"
Repository = "https://github.com/auth-platform/python-sdk"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auth_platform_sdk"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
show_error_codes = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # Pylint
    "RUF",    # Ruff-specific
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (security)
]
ignore = ["PLR0913", "PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["auth_platform_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/auth_platform_sdk"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
